requirements.txt
streamlit==1.48.0
PyPDF2==3.0.1
python-docx==1.2.0
yfinance==0.2.65
scikit-learn==1.7.1
pandas==2.3.1
numpy==2.3.2
nltk==3.9.1
matplotlib==3.10.5
plotly==6.2.0
